+++ app/src/main/java/de/passwordvault/backend/security/SecurityException.java
@@ -5,17 +5,17 @@ package de.passwordvault.backend.security;
  * Class models an {@linkplain Exception} which can be thrown on security related errors.
  *
  * @author  Christian-2003
- * @version 2.0.0
+ * @version 2.1.0
  */
-public class GenericSecurityException extends Exception {
+public class SecurityException extends Exception {
 
     /**
-     * Constructor instantiates a new {@link GenericSecurityException} that can be thrown whenever
+     * Constructor instantiates a new {@link SecurityException} that can be thrown whenever
      * security exceptions occur.
      *
      * @param message   Message to be delivered with the exception.
      */
-    public GenericSecurityException(String message) {
+    public SecurityException(String message) {
         super(message);
     }